in_opentelemetry: logs: split logs logic in a different interface
