+++ plugins/in_opentelemetry/CMakeLists.txt
@@ -6,6 +6,7 @@ set(src
   http_conn.c
   opentelemetry.c
   opentelemetry_prot.c
+  opentelemetry_logs.c
   opentelemetry_traces.c
   opentelemetry_config.c
   opentelemetry_utils.c
